+++ src/modules/pos.ts
@@ -29,6 +29,7 @@ const posFactory = {
             method: 'get',
             url: '/consumers/{consumer_id}/pos/locations',
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getOrders(
@@ -40,6 +41,7 @@ const posFactory = {
             method: 'get',
             url: '/consumers/{consumer_id}/pos/orders',
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getCustomers(
@@ -51,6 +53,7 @@ const posFactory = {
             method: 'get',
             url: '/consumers/{consumer_id}/pos/customers',
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getOrder(
@@ -61,6 +64,7 @@ const posFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/pos/orders/${orderId}`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getCustomer(
@@ -71,6 +75,7 @@ const posFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/pos/customers/${customerId}`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     createCustomer(
@@ -82,6 +87,7 @@ const posFactory = {
             url: `/consumers/{consumer_id}/pos/customers`,
             body: customer,
             clientRequestId: options?.clientRequestId,
+            datalayer: options?.datalayer,
         };
     },
     getPaymentMethods(
@@ -93,6 +99,7 @@ const posFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/pos/payment-methods`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getProductCategories(
@@ -104,6 +111,7 @@ const posFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/pos/product-categories`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getProducts(
@@ -115,6 +123,7 @@ const posFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/pos/products`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getSales(
@@ -126,6 +135,7 @@ const posFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/pos/sales`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getClosure(
@@ -138,6 +148,7 @@ const posFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/pos/closures/${date}`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getPayments(
@@ -149,6 +160,7 @@ const posFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/pos/payments`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     updateOrder(
@@ -161,6 +173,7 @@ const posFactory = {
             url: `/consumers/{consumer_id}/pos/orders/${orderId}`,
             body: order,
             clientRequestId: options?.clientRequestId,
+            datalayer: options?.datalayer,
         };
     },
     getAccountingCategories(
@@ -172,6 +185,7 @@ const posFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/pos/accounting-categories`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getObjectives(
@@ -183,6 +197,7 @@ const posFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/pos/objectives`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
 };